[DEV-173] feat: fcm 토큰 발급 유틸 함수 추가
